Leverage nlbuild-autotools-1.5.0 and allow on-demand remote synchronization of nlassert or nlunit-test from an external repository when '--with-nlassert=internal' or '--with-nlunit_test=internal' are asserted.
NLASSERT_SUBDIRS and NLUNIT_TEST_SUBDIRS are not a permanent part of DIST_SUBDIRS since we do not ever want to include them in a distribution archive; however, when they've been pulled as 'repo' modules, we do want to remove them on invocation of the 'distclean' target. Consequently, we conditionally include them in DIST_SUBDIRS on invocation of 'distclean-recursive'.